Update installation of jhbuildrc and jhbuildrc-custom...
To match the current jhbuild defaults.
Notice that one can now set XDG_CONFIG_HOME and put both there, and it
need not be in $HOME. The default is $HOME/.config. $HOME/.jhbuildrc and
$HOME/.jhbuildrc-custom are still supported and won't be disturbed if
found.

jhbuildrc will now search in order the same directory it is in,
$XDG_CONFIG_HOME if it's set or $HOME/.config if its not, and $HOME for
either jhbuildrc-custom or .jhbuildrc-custom.
If for some reason you want jhbuildrc-custom somewhere else or to have a
different name you can set $JHBUILDRC_CUSTOM. Both gtk-osx-setup.sh and
jhbuildrc will honor it.

+++ gtk-osx-setup.sh
@@ -207,16 +207,32 @@ export PATH="$PYENV_ROOT/shims:$DEVPREFIX/bin:$PYENV_INSTALL_ROOT/plugins/python
 export PYENV_ROOT
 export CFLAGS="-isysroot $SDKROOT -I$SDKROOT/usr/include"
 export PYTHON_CONFIGURE_OPTS="--enable-shared"
+export WORKON_HOME=$DEVPREFIX/share/venv
 
 $PIPENV install
 
 BASEURL="https://gitlab.gnome.org/GNOME/gtk-osx/raw/master"
 
-if test -L $HOME/.jhbuildrc ; then
-    echo "Installing jhbuild configuration..."
-    curl -ks $BASEURL/jhbuildrc-gtk-osx -o $HOME/.jhbuildrc
+config_dir=""
+if test -d "$XDG_CONFIG_HOME" ; then
+    config_dir="$XDG_CONFIG_HOME"
+elif test -d "$HOME/.config" ; then
+    config_dir = "$HOME/.config"
 fi
 
-if test ! -e $HOME/.jhbuildrc-custom ; then
+jhbuildrc_file=""
+if test -e "$config_dir/jhbuildrc" ; then
+    jhbuildrc_file="$config_dir/jhbuildrc"
+elif test -e "$HOME/.jhbuildrc" ; then
+    jhbuildrc_file="$HOME/.jhbuildrc"
+fi
+
+if -z "$jhbuildrc_file" ; then
+    echo "Installing jhbuild configuration..."
+    curl -ks $BASEURL/jhbuildrc-gtk-osx -o "$config_dir/jhbuildrc"
+fi
+
+envvar JHBUILDRC_CUSTOM "$config_dir/jhbuildrc-custom"
+if test ! -e "$JHBUILDRC_CUSTOM" -a ! -e $HOME/.jhbuildrc-custom; then
    curl -ks $BASEURL/jhbuildrc-gtk-osx-custom-example -o $HOME/.jhbuildrc-custom
 fi
